feat: refactor and if  < 0 then 0
