+++ .github/workflows/pytest.yml
@@ -18,7 +18,7 @@ jobs:
           python-version: 3.11.9
 
       - name: Install dependencies
-        run: pip install -r test_dependencies.txt
+        run: pip install -r dependencies_for_workflow.txt
 
       - name: Run Pytest tests
         run: pytest
