+++ pytest.ini
@@ -1,3 +1,2 @@
 [pytest]
 pythonpath = src
-testpaths = tests
