+++ src/main.py
@@ -1,16 +1,14 @@
 from flask import Flask, request
-import joblib
 from get_predicted_ratings import get_predicted_ratings
 from formatters.input_formatter import input_formatter
 from formatters.output_formatter import output_formatter
-from xgboost import XGBClassifier, XGBRegressor
+from xgboost import XGBRegressor
 import pandas as pd
 
 app = Flask(__name__)
 
-# load the model
+# load the regressor model
 xgb_model = XGBRegressor()
-# xgb_model = XGBClassifier()
 xgb_model.load_model('src/saved_models/xgb_model_reg.json')
 
 
@@ -24,10 +22,7 @@ def predict_ratings():
     if request.method == "POST":
         # Get the data from the request
         data = input_formatter(request.data)
-        # load scaler
-        #scaler = joblib.load('src/saved_scaler/scaler.pkl')
-        # Scale the input features
-        #scaled_data = scaler.transform(data)
+       
         # Predict ratings for all input data at once
         ratings = get_predicted_ratings(data, xgb_model)
         # Assemble the rating list with corresponding piste id
