+++ tests/data_splitter_test.py
@@ -1,7 +1,7 @@
 import pytest
 
 def test_separate_labels_success():
-		from src.training.data_splitter import separate_labels
+		from training.data_splitter import separate_labels
 		data = [
 				{"points": 1, "text": "This is a test"},
 				{"points": 2, "text": "This is a test"},
@@ -14,7 +14,7 @@ def test_separate_labels_success():
 		assert labels == [1, 2, 3, 4, 5]
 	
 def test_separate_labels_failure():
-		from src.training.data_splitter import separate_labels
+		from training.data_splitter import separate_labels
 		data = [
 				{"points": 1, "text": "This is a test"},
 				{"points": 2, "text": "This is a test"},
@@ -23,7 +23,7 @@ def test_separate_labels_failure():
 				{"points": 5.0, "text": "This is a test"}
 		]
 		with pytest.raises(ValueError) as err:
-				new_data, labels = separate_labels(data)
+				separate_labels(data)
 				assert str(err.value) == 'Label must be an integer'
 		
 		data = [
@@ -34,7 +34,7 @@ def test_separate_labels_failure():
 				{"points": 5, "text": "This is a test"}
 		]
 		with pytest.raises(KeyError) as err:
-				new_data, labels = separate_labels(data)
+				separate_labels(data)
 				assert str(err.value) == "points"
 		
 		data = [
@@ -45,6 +45,6 @@ def test_separate_labels_failure():
 				{"points": "5", "text": "This is a test"}
 		]
 		with pytest.raises(ValueError) as err:
-				new_data, labels = separate_labels(data)
+				separate_labels(data)
 				assert str(err.value) == 'Label must be an integer'
